refactor(typeahead): move wrapper and token ownership from base to consumers
XDSBaseTypeahead is now a pure combobox engine — just the input, search,
keyboard navigation, and dropdown. No wrapper div, no border styling,
no token rendering, no edit mode, no isEmbedded flag.

Consumers provide their own wrapper and pass anchorRef for dropdown
positioning:

- XDSTypeahead: owns wrapper div with borders, status styles, padding,
  token with spacing compensation, edit mode (click to edit, blur to
  restore, escape to cancel)
- XDSTokenizer: already owned its wrapper, now passes wrapperRef as
  anchorRef instead of using isEmbedded

Removed from base: isEmbedded, wrapperXStyle, statusType, startContent,
hasClear, size, wrapper/token/status styles.

Added to base: anchorRef (RefObject) for dropdown positioning. Falls
back to the input element itself if not provided.
